Eric Sunshine fixed the old renderer's gl3d, soft3d, and null3d plugins so that they once again build successfully following Anders Stenberg's rewrite of the shader variable system.  Also eliminated some related compilation warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@16904 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csgfx/shadervarcontext.cpp
@@ -61,7 +61,7 @@ void csShaderVariableContext::PushVariables
   for (int i=0; i<variables.Length (); ++i)
   {
     csStringID name = variables[i]->GetName ();
-    if (stacks.Length ()<=name)
+    if (stacks.Length () <= (int)name)
       stacks.SetLength (name+1);
     stacks[name].Push (variables[i]);
   }
